pip_requirements.txt
Babel==0.9.6
BabelDjango==0.2.2
Django==1.3.1
South==0.7.3
django-extensions==0.8
django-mptt==0.5.2
django-reversion==1.5.1
nosmsd==0.14
wsgiref==0.1.2
MySQL-python==1.2.3
xlwt==0.7.4